Move fence API to new header
