+++ include/distr_queue.h
@@ -3,111 +3,22 @@
 #include <memory>
 #include <type_traits>
 
-#include "accessor.h"
-#include "buffer_manager.h"
 #include "device_queue.h"
-#include "host_object.h"
 #include "runtime.h"
 #include "task_manager.h"
 
 namespace celerity {
-class distr_queue;
-}
+namespace detail {
 
-namespace celerity::detail {
-template <typename DataT, int Dims>
-class buffer_fence_promise;
-}
+	class distr_queue_tracker {
+	  public:
+		~distr_queue_tracker() { runtime::get_instance().shutdown(); }
+	};
 
-namespace celerity::experimental {
+	template <typename CGF>
+	constexpr bool is_safe_cgf = std::is_standard_layout<CGF>::value;
 
-template <typename T, int Dims>
-class buffer_snapshot {
-  public:
-	buffer_snapshot() : m_sr({}, detail::zero_range) {}
-
-	explicit operator bool() const { return m_data != nullptr; }
-
-	range<Dims> get_offset() const { return m_sr.offset; }
-
-	range<Dims> get_range() const { return m_sr.range; }
-
-	subrange<Dims> get_subrange() const { return m_sr; }
-
-	const T* get_data() const { return m_data.get(); }
-
-	std::unique_ptr<T[]> into_data() && { return std::move(m_data); }
-
-	inline const T& operator[](id<Dims> index) const { return m_data[detail::get_linear_index(m_sr.range, index)]; }
-
-	inline detail::subscript_result_t<Dims, const buffer_snapshot> operator[](size_t index) const { return detail::subscript<Dims>(*this, index); }
-
-	friend bool operator==(const buffer_snapshot& lhs, const buffer_snapshot& rhs) { return lhs.m_sr == rhs.m_sr && lhs.m_data == rhs.m_data; }
-
-	friend bool operator!=(const buffer_snapshot& lhs, const buffer_snapshot& rhs) { return !operator==(lhs, rhs); }
-
-  private:
-	template <typename U, int Dims2>
-	friend class detail::buffer_fence_promise;
-
-	subrange<Dims> m_sr;
-	std::unique_ptr<T[]> m_data; // cannot use std::vector here because of vector<bool> m(
-
-	explicit buffer_snapshot(subrange<Dims> sr, std::unique_ptr<T[]> data) : m_sr(sr), m_data(std::move(data)) {}
-};
-
-} // namespace celerity::experimental
-
-namespace celerity::detail {
-
-template <typename T>
-class host_object_fence_promise : public detail::fence_promise {
-  public:
-	explicit host_object_fence_promise(const experimental::host_object<T>& obj) : m_host_object(obj) {}
-
-	std::future<T> get_future() { return m_promise.get_future(); }
-
-	void fulfill() override { m_promise.set_value(std::as_const(detail::get_host_object_instance(m_host_object))); }
-
-  private:
-	experimental::host_object<T> m_host_object;
-	std::promise<T> m_promise;
-};
-
-template <typename DataT, int Dims>
-class buffer_fence_promise : public detail::fence_promise {
-  public:
-	explicit buffer_fence_promise(const buffer<DataT, Dims>& buf, const subrange<Dims>& sr) : m_buffer(buf), m_subrange(sr) {}
-
-	std::future<experimental::buffer_snapshot<DataT, Dims>> get_future() { return m_promise.get_future(); }
-
-	void fulfill() override {
-		const auto access_info = runtime::get_instance().get_buffer_manager().get_host_buffer<DataT, Dims>(
-		    get_buffer_id(m_buffer), access_mode::read, range_cast<3>(m_subrange.range), id_cast<3>(m_subrange.offset));
-		assert((access_info.offset <= m_subrange.offset) == id_cast<Dims>(id<3>(true, true, true)));
-		auto data = std::make_unique<DataT[]>(m_subrange.range.size());
-		memcpy_strided(access_info.buffer.get_pointer(), data.get(), sizeof(DataT), access_info.buffer.get_range(), m_subrange.offset - access_info.offset,
-		    m_subrange.range, {}, m_subrange.range);
-		m_promise.set_value(experimental::buffer_snapshot<DataT, Dims>(m_subrange, std::move(data)));
-	}
-
-  private:
-	buffer<DataT, Dims> m_buffer;
-	subrange<Dims> m_subrange;
-	std::promise<experimental::buffer_snapshot<DataT, Dims>> m_promise;
-};
-
-class distr_queue_tracker {
-  public:
-	~distr_queue_tracker() { runtime::get_instance().shutdown(); }
-};
-
-template <typename CGF>
-constexpr bool is_safe_cgf = std::is_standard_layout<CGF>::value;
-
-} // namespace celerity::detail
-
-namespace celerity {
+} // namespace detail
 
 struct allow_by_ref_t {};
 
@@ -188,33 +99,3 @@ class distr_queue {
 };
 
 } // namespace celerity
-
-namespace celerity::experimental {
-
-template <typename T>
-std::future<T> fence(celerity::distr_queue& q, const experimental::host_object<T>& obj) {
-	detail::side_effect_map side_effects;
-	side_effects.add_side_effect(detail::get_host_object_id(obj), experimental::side_effect_order::sequential);
-	auto promise = std::make_unique<detail::host_object_fence_promise<T>>(obj);
-	auto future = promise->get_future();
-	detail::runtime::get_instance().get_task_manager().generate_fence_task({}, std::move(side_effects), std::move(promise));
-	return future;
-}
-
-template <typename DataT, int Dims>
-std::future<buffer_snapshot<DataT, Dims>> fence(celerity::distr_queue& q, const buffer<DataT, Dims>& buf, const subrange<Dims>& sr) {
-	detail::buffer_access_map access_map;
-	access_map.add_access(detail::get_buffer_id(buf),
-	    std::make_unique<detail::range_mapper<Dims, celerity::access::fixed<Dims>>>(celerity::access::fixed<Dims>(sr), access_mode::read, buf.get_range()));
-	auto promise = std::make_unique<detail::buffer_fence_promise<DataT, Dims>>(buf, sr);
-	auto future = promise->get_future();
-	detail::runtime::get_instance().get_task_manager().generate_fence_task(std::move(access_map), {}, std::move(promise));
-	return future;
-}
-
-template <typename DataT, int Dims>
-std::future<buffer_snapshot<DataT, Dims>> fence(celerity::distr_queue& q, const buffer<DataT, Dims>& buf) {
-	return fence(q, buf, {{}, buf.get_range()});
-}
-
-} // namespace celerity::experimental
